pyproject.toml
[tool.poetry]
name = "social-plugin"
version = "0.1.0"
description = ""
authors = ["itaim <itai@rolebotics.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.92.0"
uvicorn = "^0.20.0"
authlib= "^1.2.0"
auth0-python= "^4.1.0"
cryptography = "^40.0.1"
markdown2 = "^2.4.8"
itsdangerous= "^2.1.2"
openai = "^0.27.2"
langchain = "^0.0.135"
requests-oauthlib = "^1.3.1"
python-dotenv = "^0.21.1"
pydantic = "^1.10.5"
tenacity = "^8.2.1"
tiktoken = "^0.2.0"
numpy = "^1.24.2"
docx2txt = "^0.8"
PyPDF2 = "^3.0.1"
python-pptx = "^0.6.21"
python-multipart = "^0.0.6"
arrow = "^1.2.3"
loguru = "^0.6.0"
PyJWT = "^2.6.0"
google-auth = "^2.17.2"
duckduckgo-search = "^2.8.5"
google-api-python-client = "^2.84.0"
beautifulsoup4 = "^4.12.1"
pinecone-client = "^2.2.1"
python-slugify = "^8.0.1"
tweepy = "^4.13.0"
httpx = "^0.23.3"

[tool.poetry.scripts]
start = "rolebotics.main:start"

[tool.poetry.group.dev.dependencies]

pytest = "^7.2.1"
pytest-cov = "^4.0.0"
pytest-asyncio = "^0.20.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
